requirements.txt
attrs==20.3.0
queuelib==1.5.0
requests==2.25.1
urllib3==1.26.2
w3lib==1.22.0
ujson==5.4.0
jproperties==2.1.1
beautifulsoup4==4.12.3
faiss-cpu==1.8.0.post1
numpy==1.26.4
python-dotenv==1.0.1
click==8.1.7